requirements.txt
# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
moto>=4.2.13  # For mocking AWS services in tests

# Core dependencies
pydantic>=2.9.2

# LangChain and core dependencies
langchain==0.3.18
langchain-openai==0.3.14
langchain-core==0.3.56
langchain-community==0.3.17
langchain-google-genai==2.0.8
langchain-huggingface==0.1.2
langchain-aws==0.2.15
langchain-ollama==0.2.3
langchain_xai==0.2.0
langchain_experimental==0.3.4

# LLM Providers
openai==1.76.2
anthropic==0.8.0
boto3>=1.35.74  # Required for Amazon Bedrock
ollama>=0.4.4,<1

# Database
neo4j>=5.17.0  # Neo4j Python driver
boto3>=1.35.74  # AWS SDK for Python
dynamodb-json==1.3.0  # For handling DynamoDB JSON formats
chromadb>=0.4.18  # Vector database for embeddings

# Document processing
pypdf>=4.0.0  # For PDF file handling
pdf2image>=1.17.0  # For converting PDF to images

# Attachment handling
imageio[ffmpeg]==2.37.0
pillow==11.1.0
ffmpeg==1.4.0

# Additional dependencies that might be needed
python-dotenv==1.0.0  # For environment variable management
requests==2.32.2      # Required for API calls

# Captioning + Embedding + Vector DB
torch
transformers
Pillow
chromadb
tqdm
pymupdf
# NLP and Visualization
nltk>=3.9.1  # Natural Language Toolkit (includes modules like tokenize and corpus)
matplotlib>=3.10.0  # For plotting and visualization

openai-whisper